refactor: app 레이아웃에서 헤더 컴포넌트 보여주도록 변경
