+++ src/pages/ProductPage.tsx
@@ -1,19 +1,17 @@
-import LoadingSpinner from "@/components/common/LoadingSpinner";
-import TheHeader from "@/components/layout/TheHeader";
 import ProductDescriptionSection from "@/components/products/ProductDescriptionSection";
 import ProductDetailSection from "@/components/products/ProductDetailSection";
 import ProductInfoSection from "@/components/products/ProductInfoSection";
 import ProductOrderSection from "@/components/products/ProductOrderSection";
 import ProductReviewSection from "@/components/products/ProductReviewSection";
 import withUser from "@/hoc/withUser";
-import useProductsQueries from "@/hooks/useProductsQueries";
+import useProductsQueries from "@/hooks/api/useProductsQueries";
 import styled from "@emotion/styled";
 import { useState } from "react";
 import { useParams } from "react-router";
 
 const ProductPage = () => {
   const { id } = useParams<{ id: string }>();
-  const { products, productsWish, productsDetail, productsReview, isPending } =
+  const { products, productsWish, productsDetail, productsReview } =
     useProductsQueries({ id: id ?? "" });
 
   const [tab, setTab] = useState("상품설명");
@@ -21,59 +19,46 @@ const ProductPage = () => {
     setTab(tabName);
   };
 
-  if (
-    !products ||
-    !productsWish ||
-    !productsDetail ||
-    !productsReview ||
-    isPending
-  ) {
-    return <LoadingSpinner height="266px" />;
-  }
-
   return (
-    <>
-      <TheHeader />
-      <Main>
-        <ProductInfoSection
-          name={products.name}
-          imageURL={products.imageURL}
-          price={products.price.sellingPrice}
-          brandName={products.brandInfo.name}
-          brandImageURL={products.brandInfo.imageURL}
-        />
-        <Tab>
-          <TabButton
-            onClick={() => selectTab("상품설명")}
-            selected={tab === "상품설명"}
-          >
-            상품설명
-          </TabButton>
-          <TabButton
-            onClick={() => selectTab("상품후기")}
-            selected={tab === "상품후기"}
-          >
-            상품후기
-          </TabButton>
-          <TabButton
-            onClick={() => selectTab("상세정보")}
-            selected={tab === "상세정보"}
-          >
-            상세정보
-          </TabButton>
-        </Tab>
-        {tab === "상품설명" && (
-          <ProductDescriptionSection description={productsDetail.description} />
-        )}
-        {tab === "상품후기" && (
-          <ProductReviewSection reviews={productsReview.reviews} />
-        )}
-        {tab === "상세정보" && (
-          <ProductDetailSection announcements={productsDetail.announcements} />
-        )}
-      </Main>
+    <Main>
+      <ProductInfoSection
+        name={products.name}
+        imageURL={products.imageURL}
+        price={products.price.sellingPrice}
+        brandName={products.brandInfo.name}
+        brandImageURL={products.brandInfo.imageURL}
+      />
+      <Tab>
+        <TabButton
+          onClick={() => selectTab("상품설명")}
+          selected={tab === "상품설명"}
+        >
+          상품설명
+        </TabButton>
+        <TabButton
+          onClick={() => selectTab("상품후기")}
+          selected={tab === "상품후기"}
+        >
+          상품후기
+        </TabButton>
+        <TabButton
+          onClick={() => selectTab("상세정보")}
+          selected={tab === "상세정보"}
+        >
+          상세정보
+        </TabButton>
+      </Tab>
+      {tab === "상품설명" && (
+        <ProductDescriptionSection description={productsDetail.description} />
+      )}
+      {tab === "상품후기" && (
+        <ProductReviewSection reviews={productsReview.reviews} />
+      )}
+      {tab === "상세정보" && (
+        <ProductDetailSection announcements={productsDetail.announcements} />
+      )}
       <ProductOrderSection {...productsWish} productId={id} />
-    </>
+    </Main>
   );
 };
